chore: improve vertexai example
